REFACTOR: move merch code to Content.tsx, use Next.js Image component, flatten carousel dots
